kbatch-server/requirements.txt
django
django-environ
django-rest-framework
jupyterhub
kubernetes
escapism
gunicorn
whitenoise